+++ svd-rs/CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-- Bump MSRV to 1.58.0
+## [v0.14.3] - 2023-04-04
 
+- Bump MSRV to 1.58.0
 - `array::names` returns custom names if specified
 - add `Description` trait and `array::descriptions`
+- add write constraint range check
 
 ## [v0.14.2] - 2023-04-04
 
@@ -86,7 +88,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Previous versions in common [changelog](../CHANGELOG.md).
 
-[Unreleased]: https://github.com/rust-embedded/svd/compare/svd-rs-v0.14.2...HEAD
+[Unreleased]: https://github.com/rust-embedded/svd/compare/svd-rs-v0.14.3...HEAD
+[v0.14.3]: https://github.com/rust-embedded/svd/compare/svd-parser-v0.14.2...svd-rs-v0.14.3
 [v0.14.2]: https://github.com/rust-embedded/svd/compare/svd-encoder-v0.14.2...svd-rs-v0.14.2
 [v0.14.1]: https://github.com/rust-embedded/svd/compare/v0.14.0...svd-rs-v0.14.1
 [v0.14.0]: https://github.com/rust-embedded/svd/compare/svd-rs-v0.13.2...v0.14.0